# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/sales": {
      "get": {
        "operationId": "SalesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "CONFIRMED",
                "CANCELED",
                "REFUNDED",
                "PARTIALLY_REFUNDED"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "orderId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Vendas",
        "tags": [
          "sales"
        ]
      },
      "post": {
        "operationId": "SalesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branchId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "customerId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "orderId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "qty": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "maximum": 10000
                        },
                        "unitPriceCents": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1000000000
                        },
                        "discountCents": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 1000000000,
                          "default": 0
                        }
                      },
                      "required": [
                        "productId",
                        "qty"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 200
                  },
                  "discountCents": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000000,
                    "default": 0
                  }
                },
                "required": [
                  "branchId",
                  "items"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cria venda (rascunho) com itens e descontos autorizados",
        "tags": [
          "sales"
        ]
      }
    },
    "/api/v1/sales/{id}": {
      "get": {
        "operationId": "SalesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhe da venda com pagamentos",
        "tags": [
          "sales"
        ]
      }
    },
    "/api/v1/sales/{id}/confirm": {
      "post": {
        "operationId": "SalesController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cashSessionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "payments": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string",
                          "enum": [
                            "CASH",
                            "PIX",
                            "CREDIT_CARD",
                            "DEBIT_CARD",
                            "BANK_TRANSFER",
                            "OTHER"
                          ]
                        },
                        "amountCents": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "maximum": 1000000000
                        },
                        "tenderedCents": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "maximum": 1000000000
                        },
                        "externalReference": {
                          "type": "string",
                          "maxLength": 120
                        }
                      },
                      "required": [
                        "method",
                        "amountCents"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 10
                  }
                },
                "required": [
                  "cashSessionId",
                  "payments"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirma venda: baixa estoque, registra pagamentos (único/misto) e caixa",
        "tags": [
          "sales"
        ]
      }
    },
    "/api/v1/sales/{id}/cancel": {
      "post": {
        "operationId": "SalesController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  }
                },
                "required": [
                  "reason"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancela venda em rascunho",
        "tags": [
          "sales"
        ]
      }
    },
    "/api/v1/sales/{id}/refund": {
      "post": {
        "operationId": "SalesController_refund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "cashSessionId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "method": {
                    "type": "string",
                    "enum": [
                      "CASH",
                      "PIX",
                      "CREDIT_CARD",
                      "DEBIT_CARD",
                      "BANK_TRANSFER",
                      "OTHER"
                    ]
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "saleItemId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "qty": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0
                        }
                      },
                      "required": [
                        "saleItemId",
                        "qty"
                      ],
                      "additionalProperties": false
                    },
                    "minItems": 1,
                    "maxItems": 200
                  },
                  "returnToStock": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "reason",
                  "cashSessionId",
                  "method",
                  "items"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Estorno/devolução de itens",
        "tags": [
          "sales"
        ]
      }
    },
    "/api/v1/sales/{id}/receipt": {
      "get": {
        "operationId": "SalesController_receipt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Comprovante NÃO FISCAL (PDF térmico ou A4)",
        "tags": [
          "sales"
        ]
      }
    },
...
  },
  "info": {
    "title": "OrdemCerta API",
    "description": "API REST v1 — gestão de assistências técnicas. Erros: {code,message,details,requestId}.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}